+++ src/main/java/wooteco/subway/admin/domain/graph/jgraph/SubwayJGraph.java
@@ -1,20 +1,23 @@
-package wooteco.subway.admin.domain.graph;
+package wooteco.subway.admin.domain.graph.jgraph;
 
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.WeightedMultigraph;
 import wooteco.subway.admin.domain.Edge;
+import wooteco.subway.admin.domain.graph.Graph;
+import wooteco.subway.admin.domain.graph.SubwayPath;
+import wooteco.subway.admin.domain.graph.WeightEdge;
 import wooteco.subway.admin.exception.IllegalPathRequestException;
 
 import java.util.Objects;
 import java.util.Set;
 import java.util.function.Function;
 
-public class SubwayGraph implements Graph {
+public class SubwayJGraph implements Graph {
     private WeightedMultigraph<Long, WeightEdge> graph;
 
-    public SubwayGraph(Set<Edge> edges, Function<Edge, Integer> edgeIntegerFunction) {
-        WeightedMultigraph<Long, WeightEdge> graph = new WeightedMultigraph<>(SubwayWeightEdge.class);
+    public SubwayJGraph(Set<Edge> edges, Function<Edge, Integer> edgeIntegerFunction) {
+        WeightedMultigraph<Long, WeightEdge> graph = new WeightedMultigraph<>(SubwayJWeightEdge.class);
         for (Edge edge : edges) {
             addEdge(graph, edge, edgeIntegerFunction);
         }
@@ -25,7 +28,7 @@ public class SubwayGraph implements Graph {
         graph.addVertex(edge.getStationId());
         if (edge.isNotFirst()) {
             graph.addVertex(edge.getPreStationId());
-            graph.addEdge(edge.getPreStationId(), edge.getStationId(), new SubwayWeightEdge(edge, edgeIntegerFunction));
+            graph.addEdge(edge.getPreStationId(), edge.getStationId(), new SubwayJWeightEdge(edge, edgeIntegerFunction));
         }
     }
 
@@ -61,7 +64,7 @@ public class SubwayGraph implements Graph {
     public boolean equals(final Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        final SubwayGraph that = (SubwayGraph) o;
+        final SubwayJGraph that = (SubwayJGraph) o;
         return Objects.equals(graph, that.graph);
     }
